+++ packages/@ember/-internals/runtime/lib/mixins/registry_proxy-internal.ts
