+++ packages/@ember/-internals/runtime/tests/mixins/container_proxy_test.js
@@ -1,6 +1,6 @@
 import { getOwner } from '@ember/-internals/owner';
 import { Container, Registry } from '@ember/-internals/container';
-import ContainerProxy from '../../lib/mixins/container_proxy';
+import ContainerProxy from '../../lib/mixins/container_proxy-internal';
 import EmberObject from '@ember/object';
 import { run, schedule } from '@ember/runloop';
 import { moduleFor, AbstractTestCase } from 'internal-test-helpers';
