+++ packages/@ember/engine/index.ts
@@ -14,7 +14,7 @@ import type { EngineInstanceOptions } from '@ember/engine/instance';
 import EngineInstance from '@ember/engine/instance';
 import { RoutingService } from '@ember/routing/-internals';
 import { setupEngineRegistry } from '@ember/-internals/glimmer/lib/setup-registry';
-import RegistryProxyMixin from '@ember/-internals/runtime/lib/mixins/registry_proxy';
+import RegistryProxyMixin from '@ember/-internals/runtime/lib/mixins/registry_proxy-internal';
 import { StrictResolver } from './lib/strict-resolver';
 
 function props(obj: object) {
