+++ packages/@ember/engine/instance.ts
@@ -8,8 +8,8 @@ import { assert } from '@ember/debug';
 import { default as Registry, privatize as P } from '@ember/-internals/container/lib/registry';
 import { guidFor } from '@ember/-internals/utils/lib/guid';
 import { ENGINE_PARENT, getEngineParent, setEngineParent } from './parent';
-import ContainerProxyMixin from '@ember/-internals/runtime/lib/mixins/container_proxy';
-import RegistryProxyMixin from '@ember/-internals/runtime/lib/mixins/registry_proxy';
+import ContainerProxyMixin from '@ember/-internals/runtime/lib/mixins/container_proxy-internal';
+import RegistryProxyMixin from '@ember/-internals/runtime/lib/mixins/registry_proxy-internal';
 import type { InternalOwner } from '@ember/-internals/owner';
 import type Owner from '@ember/-internals/owner';
 import { type FullName, isFactory } from '@ember/-internals/owner';
